+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   * [FEATURE] Experimental `ext-gridnav` implents key navigation for tables.
     Refactored keyboard handling. Keydown handlers are now bound to the container instead of document  
     (Co-work with Koloto)
+  * [FEATURE] Allow to return 'preventNav' in keydown event to prevent withput blocking keys in embedded input controls.
+  * [FEATURE] new method `node.navigate()` to support custom keyboard handlers
   * [FEATURE] Refactored CSS style to use a common LESS template
   * [FEATURE] Improvement of lazy load errors handling (Koloto, issue #87)
   * [BUG] Fixed BACKSPACE on top-level nodes
@@ -12,11 +14,12 @@
   * Improved table render speed by 15%
   * `grunt dev` combines `grunt server` + `grunt watch` (trigger jshint and 
     less on save)
+  * [FEATURE] Allow to pass metadata with `source` on initialization
 
 
 # 2.0.0-4
 
-  * Fixes
+  * Misc. fixes
 
 
 # 2.0.0-3